Added preview gif and Fixed toaster
